Force cursor to line beginning if SSH_CONNECTION
This commit fixes a bug where PROMPT would be drawn twice. This could
happen if the PROMPT was rendered at the bottom of the terminal. Due to
the escape codes \e7 and \e8 capturing the absolute position of the
terminal (they speak with the terminal, not the shell), the would move
the cursor to the bottom of the terminal viewport on restore, thus
drawing the PROMPT on two consecutive lines.